Add: add minimum/maximum thinking time option

+++ src/command/executor.cc
@@ -148,10 +148,10 @@ void Executor::setConfig(const commands::BoolConfig* Config) {
 
 void Executor::setConfig(const commands::IntegerConfig* Config) {
     if (Config->configurable() == commands::Configurable::MinimumThinkingTime) {
-        CManager.setMinimumThinkinTimeMilliSeconds(Config->value());
+        CManager.setMinimumThinkinTimeMilliSeconds((uint32_t)Config->value());
     } else if (Config->configurable() ==
                commands::Configurable::MaximumThinkingTime) {
-        CManager.setMaximumThinkinTimeMilliSeconds(Config->value());
+        CManager.setMaximumThinkinTimeMilliSeconds((uint32_t)Config->value());
     } else if (Config->configurable() == commands::Configurable::MaxPly) {
         StateConfig->MaxPly = (uint16_t)Config->value();
     } else if (Config->configurable() == commands::Configurable::NumGPUs) {
